e2e: add Babylon node/client to e2e test manager (#7)
