+++ Makefile
@@ -8,6 +8,7 @@ TOOLS_DIR := tools
 
 BTCD_PKG := github.com/btcsuite/btcd
 BTCDW_PKG := github.com/btcsuite/btcwallet
+BABYLON_PKG := github.com/babylonchain/babylon/cmd/babylond
 
 GO_BIN := ${GOPATH}/bin
 BTCD_BIN := $(GO_BIN)/btcd
@@ -43,7 +44,7 @@ test:
 	go test ./...
 
 test-e2e:
-	cd $(TOOLS_DIR); go install -trimpath $(BTCD_PKG); go install -trimpath $(BTCDW_PKG)
+	cd $(TOOLS_DIR); go install -trimpath $(BTCD_PKG); go install -trimpath $(BTCDW_PKG); go install -trimpath $(BABYLON_PKG);
 	go test -mod=readonly -timeout=25m -v $(PACKAGES_E2E) -count=1 --tags=e2e
 
 build-docker:
