+++ e2etest/log.go
@@ -1,0 +1,7 @@
+package e2etest
+
+import (
+	vlog "github.com/babylonchain/vigilante/log"
+)
+
+var log = vlog.Logger.WithField("module", "e2etest")
