+++ tools/tools.go
@@ -4,6 +4,7 @@
 package vigilante
 
 import (
+	_ "github.com/babylonchain/babylon/cmd/babylond"
 	_ "github.com/btcsuite/btcd"
 	_ "github.com/btcsuite/btcwallet"
 )
